% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.norm.R
\name{frobenius.norm}
\alias{frobenius.norm}
\title{Compute the Frobenius Norm of a Matrix}
\usage{
frobenius.norm(M)
}
\arguments{
\item{M}{A numeric matrix.}
}
\value{
A non-negative scalar representing the Frobenius norm of \code{M}.
}
\description{
Calculates the Frobenius norm of a matrix, defined as the square root of the sum of squares of all its entries. This is the standard Euclidean norm when the matrix is treated as a vector.
}
\examples{
M <- matrix(1:6, nrow = 2)
frobenius.norm(M)

}
