% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGulPC.R
\name{DGulPC}
\alias{DGulPC}
\title{Distributed general unilateral loading principal component}
\usage{
DGulPC(data, m, n1, K)
}
\arguments{
\item{data}{is a total data set}

\item{m}{is the number of principal component}

\item{n1}{is  the length of each data subset}

\item{K}{is the number of nodes}
}
\value{
AU1,AU2,DU3,Shat
}
\description{
Distributed general unilateral loading principal component
}
\examples{
library(LFM)
data_from_package <- Wine
data_a <- Wine
DGulPC(data_a,m=3,n1=128,K=2)
}
