% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPC.R
\name{DPC}
\alias{DPC}
\title{Distributed principal component}
\usage{
DPC(data, m, n1, K)
}
\arguments{
\item{data}{is a total data set}

\item{m}{is the number of principal component}

\item{n1}{is  the length of each data subset}

\item{K}{is the number of nodes}
}
\value{
Ahat,Dhat,Sigmahathat
}
\description{
Distributed principal component
}
\examples{
library(LFM)
data_from_package <- Wine
data_a <- Wine
DPC(data_a,m=3,n1=128,K=2)
}
