% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GulPC.R
\name{GulPC}
\alias{GulPC}
\title{General unilateral loading principal component}
\usage{
GulPC(data, m)
}
\arguments{
\item{data}{is a total data set}

\item{m}{is the number of first layer principal component}
}
\value{
AU1,AU2,DU3,SigmaUhat
}
\description{
General unilateral loading principal component
}
\examples{
library(LFM)
data_from_package <- Wine
data_a <- Wine
GulPC(data=data_a,m=5)
}
