% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.R
\name{PC}
\alias{PC}
\title{Principal component}
\usage{
PC(data, m)
}
\arguments{
\item{data}{is a total data set}

\item{m}{is the number of principal component}
}
\value{
Ahat, Dhat, Sigmahat
}
\description{
Principal component
}
\examples{
library(LFM)
data_from_package <- Wine
data_a <- Wine
PC(data_a,m=5)
}
