% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPC.R
\name{PPC}
\alias{PPC}
\title{Projection principal component}
\usage{
PPC(data, m)
}
\arguments{
\item{data}{is a total  data set}

\item{m}{is the number of principal component}
}
\value{
Apro, Dpro, Sigmahatpro
}
\description{
Projection principal component
}
\examples{
library(LFM)
data_from_package <- Wine
data_a <- Wine
PPC(data=data_a,m=5)
}
